Move plug and play to its own loop. Write code for registers
Plug and play feature loop should happen before the node starts
any other kind of UAVCAN activity. This means NodeID allocation
which is not complete yet, but has the appropriate states and
timers ready to be assembled in the switch statement.

Code for registers was written, but it isn't tested yet.