Fix list comparison
Comparing a Python list with a numpy list caused the problem that
identical lists weren't equal so now they are. That means that
node_id allocations are detected again by the get_info_handler
which compares a target_hw_id with the node that was allocated a
node_id recently.