Progress
There must be something that causes Sapog not to receive the
transfer sometimes and the most likely cause is a too long time
out of a transfer_id counter. I ran 7 tests and the 7th failed
because of a lost transfer. After a restart running the 7th test
and even running all 7 tests worked.